requirements.txt
selenium>=4.11.2
beautifulsoup4>=4.12.2
pandas>=2.0.3
webdriver-manager>=4.0.1
openpyxl>=3.1.2
flask>=2.0.1
flask-cors>=3.0.10
psycopg2-binary>=2.9.7